+++ apps/guard/app/SideBar.tsx
@@ -1,5 +1,6 @@
 import Link from 'next/link';
 import { usePathname, useRouter } from 'next/navigation';
+import { useMemo, useState } from 'react';
 
 import {
   BitcoinCircle,
@@ -9,7 +10,13 @@ import {
   Moneybag,
   ClipboardNotes,
 } from '@rosen-bridge/icons';
-import { AppBar, AppLogo } from '@rosen-bridge/ui-kit';
+import {
+  AppBar,
+  AppLogo,
+  NavigationBar,
+  NavigationButton,
+  Version,
+} from '@rosen-bridge/ui-kit';
 
 import packageJson from '../package.json';
 import { useInfo } from './_hooks/useInfo';
@@ -22,79 +29,87 @@ export const SideBar = () => {
 
   const router = useRouter();
 
-  const { data: info, isLoading } = useInfo();
-
-  const versions = [
+  const routes = [
     {
-      title: 'Guard',
-      value: info?.versions.app,
-      important: true,
+      label: 'Dashboard',
+      path: '/',
+      icon: <Dashboard />,
     },
     {
-      title: 'UI',
-      value: packageJson.version,
+      label: 'Health',
+      path: '/health',
+      icon: <Heartbeat />,
     },
     {
-      title: 'Contract',
-      value: info?.versions.contract,
+      label: 'Events',
+      path: '/events',
+      icon: <ClipboardNotes />,
+    },
+    {
+      label: 'History',
+      path: '/history',
+      icon: <History />,
+    },
+    {
+      label: 'Assets',
+      path: '/assets',
+      icon: <BitcoinCircle />,
+    },
+    {
+      label: 'Revenues',
+      path: '/revenues',
+      icon: <Moneybag />,
     },
   ];
 
-  if (!isLoading && info?.versions.contract !== info?.versions.tokensMap) {
-    versions.push({
-      title: 'Tokens',
-      value: info?.versions.tokensMap,
-    });
-  }
+  const { data: info, isLoading } = useInfo();
+
+  const sub = useMemo(() => {
+    const result = [
+      {
+        label: 'UI',
+        value: packageJson.version,
+      },
+      {
+        label: 'Contract',
+        value: info?.versions.contract,
+      },
+    ];
+    if (!isLoading && info?.versions.contract !== info?.versions.tokensMap) {
+      result.push({
+        label: 'Tokens',
+        value: info?.versions.tokensMap,
+      });
+    }
+    return result;
+  }, [info, isLoading]);
 
   return (
     <AppBar
       logo={
         <Link href="/">
-          <AppLogo darkLogoPath="/dark.png" lightLogoPath="/light.png" />
+          <AppLogo
+            darkLogoPath="/logo-dark-desktop.png"
+            lightLogoPath="/logo-light-desktop.png"
+            darkLogoMobilePath="/logo-dark-mobile.png"
+            lightLogoMobilePath="/logo-light-mobile.png"
+          />
         </Link>
       }
-      versions={versions}
-      routes={[
-        {
-          label: 'Dashboard',
-          path: '/',
-          disabled: false,
-          icon: <Dashboard />,
-        },
-        {
-          label: 'Health',
-          path: '/health',
-          disabled: false,
-          icon: <Heartbeat />,
-        },
-        {
-          label: 'Events',
-          path: '/events',
-          disabled: false,
-          icon: <ClipboardNotes />,
-        },
-        {
-          label: 'History',
-          path: '/history',
-          disabled: false,
-          icon: <History />,
-        },
-        {
-          label: 'Assets',
-          path: '/assets',
-          disabled: false,
-          icon: <BitcoinCircle />,
-        },
-        {
-          label: 'Revenues',
-          path: '/revenues',
-          disabled: false,
-          icon: <Moneybag />,
-        },
-      ]}
-      isActive={(route) => pathname === route.path}
-      onNavigate={(route) => router.push(route.path)}
+      versions={<Version label="Guard" value={info?.versions.app} sub={sub} />}
+      navigationBar={
+        <NavigationBar>
+          {routes.map((route) => (
+            <NavigationButton
+              key={route.label}
+              icon={route.icon}
+              isActive={pathname === route.path}
+              label={route.label}
+              onClick={() => router.push(route.path)}
+            />
+          ))}
+        </NavigationBar>
+      }
     />
   );
 };