+++ apps/rosen/app/_hooks/useMaxTransfer.ts
@@ -33,6 +33,8 @@ export const useMaxTransfer = () => {
   const loading = isTokenBalanceLoading || isTransitionLoading;
 
   const load = useCallback(async () => {
+    setMax(0n);
+
     const skip =
       !amount ||
       isTokenBalanceLoading ||
@@ -46,16 +48,6 @@ export const useMaxTransfer = () => {
     setError(false);
 
     try {
-      let eventData: any;
-
-      if (selectedSource.name === NETWORKS.BITCOIN) {
-        eventData = {
-          fromAddress: await selectedWallet.getAddress(),
-          toAddress: walletAddressValue,
-          toChain: targetValue,
-        };
-      }
-
       const max = await getMaxTransfer(
         selectedSource,
         {