+++ networks/ethereum/src/index.ts
@@ -1,4 +1,3 @@
 export * from './constants';
 export * from './generateTxParameters';
-export * from './types';
 export * from './utils';