+++ wallets/flint-wallet/src/flintWalletInfo.ts
@@ -1,9 +1,0 @@
-import { FlintIcon } from '@rosen-bridge/icons';
-import { WalletInfo } from '@rosen-ui/wallet-api';
-
-export const flintWalletInfo: WalletInfo = {
-  icon: FlintIcon,
-  name: 'Flint',
-  label: 'Flint',
-  link: 'https://flint-wallet.com/',
-};