+++ wallets/nautilus-wallet/src/nautilusWalletInfo.ts
@@ -1,9 +1,0 @@
-import { NautilusIcon } from '@rosen-bridge/icons';
-import { WalletInfo } from '@rosen-ui/wallet-api';
-
-export const nautilusWalletInfo: WalletInfo = {
-  icon: NautilusIcon,
-  name: 'Nautilus',
-  label: 'Nautilus',
-  link: 'https://github.com/nautls/nautilus-wallet',
-};