+++ wallets/wallet-api/src/types/cardano/address.ts
@@ -1,3 +1,0 @@
-import { Bech32String } from '../common';
-
-export type Addr = Bech32String;