+++ wallets/wallet-api/src/types/cardano/loaders.ts
@@ -1,2 +1,0 @@
-export type CardanoWasm =
-  typeof import('@emurgo/cardano-serialization-lib-nodejs/cardano_serialization_lib');