+++ wallets/wallet-api/src/types/ergo/ergoTree.ts
@@ -1,3 +1,0 @@
-import { HexString } from '../common';
-
-export type ErgoTree = HexString;