+++ wallets/xdefi-wallet/src/global.d.ts
@@ -1,7 +1,0 @@
-declare global {
-  interface Window {
-    xfi: { bitcoin: unknown };
-  }
-}
-
-export {};