+++ wallets/xdefi-wallet/src/isXdefiAvailable.ts
@@ -1,3 +1,0 @@
-export const isXdefiAvailable = (): boolean => {
-  return typeof window.xfi !== 'undefined' && !!window.xfi.bitcoin;
-};