+++ apps/guard/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @rosen-bridge/guard-app
 
+## 1.9.1
+
+### Patch Changes
+
+- Integrated the Version component in the SideBar to display version information optimized for both desktop and mobile views.
+- Updated dependencies
+  - @rosen-bridge/icons@1.0.0
+  - @rosen-bridge/ui-kit@1.7.0
+
+## 1.9.0
+
+### Minor Changes
+
+- Update the root layout to incorporate the latest changes from the UI kit, enhancing its responsiveness
+
+### Patch Changes
+
+- Fix missing dependencies and remove unused packages
+- Updated dependencies
+  - @rosen-bridge/ui-kit@1.6.0
+  - @rosen-ui/utils@0.4.2
+
 ## 1.8.0
 
 ### Minor Changes