+++ packages/shared-contexts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rosen-bridge/shared-contexts
 
+## 0.0.2
+
+### Patch Changes
+
+- Update the source code to address the issues identified by the new ESLint rules configuration
+
 ## 0.0.1
 
 ### Patch Changes